pyproject.toml
[tool.poetry]
name = "airdoctor"
version = "0.1.0"
description = "tools to audit your airbyte install"
authors = ["Michel Tricot <michel@airbyte.io>"]
license = "ELv2"
readme = "README.md"

[tool.poetry.scripts]
airdoctor = "airdoctor.entrypoint:main"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
coloredlogs = "^15.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.6"
poethepoet = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "N",    # PEP8 naming convetions
]
ignore = [
    "D103", # Missing docstring in public function
    "D104"  # Missing docstring in public package
]

[tool.poe.tasks]
#test         = "pytest --cov=my_app"
check = "ruff check"
_fix = "ruff check --fix"
_format = "ruff format"
format = ["_format", "_fix"]